+++ packages/common/src/index.ts
@@ -16,3 +16,5 @@ export * from './time';
 export * from './get-param-names';
 export * from './lazy-subject';
 export * from './check-native-async';
+export * from './is-constructor';
+export * from './padding';